Create Cia tools
CIA-hacking-tool
This script by D3D_M4N (me), this is a CIA themed hacking tool for an "operative" that makes it so you don't have to use arguments when typing in commands and I'm adding more tools besides nmap soon. This was designed for a linux machine with an internet connection and the following packages: git, nmap, w3m, if you do not have these packages, there will be an option in the menu. when you first execute there will be a blank space, and you will type in the password, "123123qq" (without the quotes). if you want to change it, you will have to look in the source code, it is on line 23, if [ "$pass" = "123123qq" ], you can change the variable 123123qq to whatever you want. I am currently making a version called CIA X that i will sell for money, it has more tools,more exploits, and it has a cooler UI.

How to run:

git clone https://github.com/user421921/CIA-hacking-tool.git

cd CIA-hacking-tool

chmod +x 'CIA Tool'

sudo ./'CIA Tool'